fix(simulator): desktop simulator builds again
Pre-existing breakage (since the macro/preset system + PicoSeqRack landed) — the
desktop sim couldn't compile ctagSoundProcessorGrooveBoxRack.cpp (was PicoSeqRack):
- add missing fake-idf stub simulator/fake-idf/esp_timer.h (esp_timer_get_time)
- add esp_get_free_heap_size / esp_get_minimum_free_heap_size to fake-idf/esp_system.h
- add heap_caps_get_minimum_free_size to fake-idf/esp_heap_caps.{h,c}
- ctagSoundProcessorGrooveBoxRack.cpp: portable pointer->int cast for the
  ESP_LOGI debug lines ((unsigned int)(uintptr_t)ptr) — works on ESP32 (32-bit)
  and 64-bit desktop; no behaviour change on device

Also serves as the build-check for the PicoSeqRack -> GrooveBoxRack rename:
the simulator now compiles every sound processor (incl. the renamed file),
links the generated factory, and runs.

+++ components/ctagSoundProcessor/ctagSoundProcessorGrooveBoxRack.cpp
@@ -903,18 +903,18 @@ void ctagSoundProcessorGrooveBoxRack::Init(std::size_t blockSize, void* blockPtr
 
     // delay
     delayBuffer_l = static_cast<float*>(heap_caps_malloc(delayBufferSizeMax * sizeof(float), MALLOC_CAP_SPIRAM));
-    ESP_LOGI("ctagSoundProcessorGrooveBoxRack", "Allocate: delayBuffer_l=0x%x", (unsigned int)delayBuffer_l);
+    ESP_LOGI("ctagSoundProcessorGrooveBoxRack", "Allocate: delayBuffer_l=0x%x", (unsigned int)(uintptr_t)delayBuffer_l);
     assert(delayBuffer_l != nullptr);
     std::fill_n(delayBuffer_l, delayBufferSizeMax, 0.f);
 
     delayBuffer_r = static_cast<float*>(heap_caps_malloc(delayBufferSizeMax * sizeof(float), MALLOC_CAP_SPIRAM));
-    ESP_LOGI("ctagSoundProcessorGrooveBoxRack", "Allocate: delayBuffer_r=0x%x", (unsigned int)delayBuffer_r);
+    ESP_LOGI("ctagSoundProcessorGrooveBoxRack", "Allocate: delayBuffer_r=0x%x", (unsigned int)(uintptr_t)delayBuffer_r);
     assert(delayBuffer_r != nullptr);
     std::fill_n(delayBuffer_r, delayBufferSizeMax, 0.f);
 
     // reverb
     reverbBuffer = static_cast<float*>(heap_caps_malloc(32768 * sizeof(float), MALLOC_CAP_SPIRAM));
-    ESP_LOGI("ctagSoundProcessorGrooveBoxRack", "Allocate: reverbBuffer=0x%x", (unsigned int)reverbBuffer);
+    ESP_LOGI("ctagSoundProcessorGrooveBoxRack", "Allocate: reverbBuffer=0x%x", (unsigned int)(uintptr_t)reverbBuffer);
     assert(reverbBuffer != nullptr);
     std::fill_n(reverbBuffer, 32768, 0.f);
